# src/main/resources/static/boat4you_ws_entities_crud.openapi.yaml
openapi: 3.0.3
info:
  title: Boat4You Web Service
  description: |-
    API definition for Boat4You Web Service. \
    Contains operations related to Users
  version: 0.0.1
servers:
  - url: https://localhost:8443
    description: Local deployment
...
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      requestBody:
        description: ''
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
  /auth/refreshToken:
    post:
      tags:
        - Auth
      summary: Issues a refresh token to authenticated user
      description: ''
      operationId: refreshToken
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
...
  /auth/updatePassword:
    post:
      tags:
        - Auth
      summary: Update User password
      description: This can only be done by the logged in user. Users can only change their own password.
      operationId: updatePassword
      requestBody:
        description: Update password of an existing User
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordBody'
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /auth/requestPasswordReset:
    post:
      tags:
        - Auth
      summary: Request password reset for a user
      description: Results in sending a password reset email. This can be done by Manager, System Administrator or an anonymous user.
      operationId: requestPasswordReset
      requestBody:
        description: Request password reset body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetBody'
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /auth/resetPassword:
    get:
      tags:
        - Auth
      summary: Check password reset validity
      description: Called when User clicks on the password reset link, so the FE knows if the password reset link is invalid.
      operationId: checkPasswordResetValidity
      parameters:
        - name: passwordResetCode
          in: query
          description: Password reset code of the password reset request
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
    post:
      tags:
        - Auth
      summary: Set a new password as part of the password reset process
      description: Set a new password as part of the password reset process. This can be done by an anonymous user.
      operationId: resetPassword
      parameters:
        - name: passwordResetCode
          in: query
          description: Password reset code of the password reset request
          required: true
          schema:
            type: string
      requestBody:
        description: Reset password body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserPasswordBody'
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get info about current user
      description: This can only be done by the logged in user. Logged users can only get information about themselves.
      operationId: getAuthenticatedUserInfo
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '404':
          description: User not found
        '500':
          description: Internal server error
  /auth/register:
    put:
      tags:
        - Auth
      summary: User registration
      description: User registration, anyone can register
      operationId: registerUser
      requestBody:
        description: ''
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /auth/register/verifyEmail:
    post:
      tags:
        - Auth
      summary: Email verification
      description: Used when registering, when user submits the verification code.
      operationId: verifyEmail
      requestBody:
        description: ''
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEmailVerificationRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
  /auth/register/resendVerificationCode/{userId}:
    post:
      tags:
        - Auth
      summary: Resend email verification code
      description: Used when registering, when user wants the email verification code to be resent.
      operationId: resendEmailVerificationCode
      parameters:
        - name: userId
          in: path
          description: Id of the user who wants a new verification code
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '403':
          description: Not authorized
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LanguageEnum:
      type: string
      enum:
        - EN
        - FR
        - DE
        - PT
        - IT
        - ES
        - HR
    CurrencyEnum:
      type: string
      enum:
        - EUR
        - USD
        - GBP
        - CAD
        - AUD
        - CHF
        - ARS
        - ZAR
        - BRL
        - NOK
        - CZK
        - DKK
        - ILS
        - SGD
        - NZD
        - SEK
    User:
      type: object
      required:
        - email
        - name
        - surname
        - roles
      properties:
        id:
          type: integer
          format: int64
          example: 10
        password:
          type: string
          description: Password in clear text
        name:
          type: string
          example: Blaž
        surname:
          type: string
          example: Slišković
        email:
          type: string
          example: bsliskov@workspace.hr
        phoneNumber:
          type: string
          description: Phone number in international format
          example: +385911002000
        language:
          type: string
          $ref: '#/components/schemas/LanguageEnum'
        currency:
          type: string
          $ref: '#/components/schemas/CurrencyEnum'
        userStatus:
          $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/BasicEntityStatus'
        inviteStatus:
          $ref: '#/components/schemas/UserInviteStatus'
        roles:
          description: List of roles assigned to the user
          type: array
          items:
            $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/UserRole'
    UserInviteStatus:
      description: Status of User invitation. Not sent in a request, just received in a response.
      type: string
      readOnly: true
      enum:
        - NOT_INVITED
        - INVITED
        - ACCEPTED
    UserRegistrationRequest:
      type: object
      description: Used when User is being registered via the public endpoint
      required:
        - name
        - surname
        - email
        - password
      properties:
        name:
          type: string
          example: Blaž
        surname:
          type: string
          example: Slišković
        email:
          type: string
          example: bsliskov@workspace.hr
        password:
          type: string
          description: Password in clear text
        phoneNumber:
          type: string
          description: Phone number in international format
          example: +385911002000
    UserEmailVerificationRequest:
      type: object
      description: Used when User is verifying their email address during registration
      required:
        - userId
        - verificationCode
      properties:
        userId:
          description: Id of the User whose email is being verified
          type: integer
          format: int64
          example: 10
        verificationCode:
          type: string
          description: Verification code in clear text
    RequestPasswordResetBody:
      type: object
      description: Either userId can be provided (if password reset is requested by Manager or System Administrator) or email address (if the request is made by anonymous user)
      properties:
        userId:
          description: Id of the User whose password needs to be reset
          type: integer
          format: int64
          example: 10
        email:
          description: Email address of the User whose password needs to be reset
          type: string
    SetUserPasswordBody:
      type: object
      required:
        - password
      properties:
        password:
          description: New password in clear text.
          type: string
          format: password
...
    UserLoginRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          description: Email of the user
          type: string
        password:
          description: The password for login in clear text
          type: string
          format: password
    TokenResponse:
      required:
        - token
        - refreshToken
        - userId
      type: object
      properties:
        token:
          description: Authentication token
          type: string
        refreshToken:
          description: Refresh token
          type: string
        userId:
          description: Id of the User the token belongs to
          type: integer
          format: int64
          example: 10
    UpdateUserPasswordBody:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          description: Old password in clear text.
          type: string
          format: password
        newPassword:
          description: New password in clear text.
          type: string
          format: password
security:
  - bearerAuth: []